Splitting the main.py file into smaller modules to improve maintainability and facilitate better control of the codebase.
The motivation behind splitting the main.py file into smaller modules was primarily to improve maintainability, making it easier to manage and update the codebase over time. Additionally, this change is expected to make it easier for ChatGPT to understand and work with the codebase, as it can now parse and comprehend smaller, more focused modules rather than a monolithic main.py file.
