+++ scripts/ai_config.py
